+++ AdventOfCode2019/Solver.cs
@@ -6,7 +6,7 @@ namespace AdventOfCode2019
 {
     public abstract class Solver
     {
-        public abstract int SolvePart1();
-        public abstract int SolvePart2();
+        public abstract long SolvePart1();
+        public abstract long SolvePart2();
     }
 }
